🚚 move editor into its own folder
